+++ docs/Changelog.rst
@@ -8,8 +8,9 @@ Changelog
 **Version: 0.0.2 -- currently unreleased**
 ''''''''''''''''''
 
-- All printing is in yellow
-- Improved regex to deal with certain version patterns
+- All printing is in red
+- Improved regex to deal with detection of certain version patterns
+- Improved padding of version strings to deal with comparison of certain version patterns
 
 **Version: 0.0.1 -- released 24 Mar 2021**
 '''''''''''''''''''''''''''''''''''''''''